+++ CHANGELOG.md
@@ -4,13 +4,11 @@
 * [x] Tools will wear out after using them
 * [] Items can be dropped from the inventory
 * [] Items can be placed on shortcut from the inventory
-* [] Broken blocks have a chance to give a random item
-* [] Blocks can be collected
-* [] Collected blocks can be placed back to the map
 * [x] i386 tar build
 
 ###### FIXED:
 * [x] Clear out some really nasty stack smashing bugs
+* [x] When the inventory opened, the selected item is not marked
 
 ###### INTERNAL:
 * [x] Logging